Extract CSR class hierarchy to new file
