+++ riscv/riscv.mk.in
@@ -31,6 +31,7 @@ riscv_hdrs = \
 	debug_rom_defines.h \
 	remote_bitbang.h \
 	jtag_dtm.h \
+	csrs.h \
 
 riscv_install_hdrs = mmio_plugin.h
